Makefile: fork/exec based unit test
This will simplify all the executables are generated
out of the single C file, e.g. inode.c, and unit test
those through the fork()/exec() call.  In a sence, it's
not a unit test but this will allow me to write lots
of system call examples, not only in those executables
but also in a test side, e.g. fork()/exec() and signales.