server: Add worker threads (multithread version)
-s command line option is added to fall back to the the original
sequential/single threaded version.

It uses semaphore, mutex, and condition variable, to implement
the thread pool.  I feel it's too complicated to do such a generic
work.  As an alternative approach, I'll try to call accpet(2) from
multiple thread to see if that works, as it will reduce the
interaction between the dispacher and the workers.  Or, just use
the epoll, but then, I can't play with the multi thread programming.