server: waitpid() inside SIGCHLD handler to cleanup the servers
Still there is a race condition that waitpid() might fail if
the other child is not exit()ed yet, and the SIGCHLD handler
miss those SIGCHLD signals.  I'll write another program, sigchld,
to take care of this behavior.  Based on all this process management
handling, it might be better to go with the multi-thread approach.

Let's see how it goes.