journal: Notify the systemd whenever there is an logs
It also catches SIGABRT and SIGHUP when gracefully terminate the
process.  This will be used by the systemd service file to restart
the journal service in case there is no log output for the
WatchdogSec period.  Here is the sample service file for your
reference.

[Service]
WatchdogSec=1m
WatchdogSignal=SIGHUP
Restart=always